[BUGFIX] Adjust pagination widget to recent changes
With #43589 the action argument of link/URI ViewHelpers is
not optional any longer.

This change adjusts the template of the provided pagination widget
to that breaking change by specifying the action argument.

Besides this adds a check to widget.link & widget.uri ViewHelpers
which throws a more meaningful exception if the action argument
was omitted.

Change-Id: I86ea27aa95ce23991924e51576389fb05680b500
Fixes: #44376
Related: #43589
Releases: master