[FEATURE] Support for redirects within widgets
Redirects and forwards did not work from within widget controllers
because there was no internal dispatch-loop.

This change implements support for redirects and forwards.
Besides this adds a controller check that throws an
exception when trying to redirect to a different controller
than the one associated with the widget ViewHelper.

Change-Id: I0c636ffd6a4152621663ca8e993be85f6b4fc2ad
Resolves: #31765
Releases: master