[BUGFIX] preset radio and checkboxes after validation errors
Radiobuttons and Checkboxes specified with name="…"
instead of property="…" are now checked again after
a validation error in a form occurred.  Before this change
the status of a checked radiobutton was lost as soon as a
validation error occurred in the form (so that the user had
to check it again).

Change-Id: I38de3369f47a53f0e98ddf5327c96985c0d3754a
Resolves: #34682
Releases: 2.1